PLATFORM-5758 | Fix ContentParser::getRevision on Move to existing page
During page move not all changes are populated from replica to master. We should make sure that the fetched revision is newest one, to prevent error of invalid page_id "owner". In future version of MW and SMW there is a lot of changes around RevisionStore, which could fix this issue, so it should be revisited on next upgrade Providing LatestRevId from title making it first to check in REPLICA_DB, and if it don't find it there it will try to request from MASTER_DB.